betalytics-scraper/requirements.txt
playwright==1.47.0
python-dotenv>=1.0.1
urllib3>=2.2.2
beautifulsoup4>=4.12.3
tqdm>=4.66.5
